Extract out CLI specific link code
