+++ relayer/src/link/relay_path.rs
@@ -1,8 +1,6 @@
 use alloc::collections::BTreeMap as HashMap;
 use alloc::collections::VecDeque;
-use std::convert::TryInto;
 use std::ops::Sub;
-use std::thread;
 use std::time::{Duration, Instant};
 
 use ibc_proto::google::protobuf::Any;
@@ -75,8 +73,8 @@ pub struct RelayPath<ChainA: ChainHandle, ChainB: ChainHandle> {
     // mostly timeout packet messages.
     // The operational data targeting the destination chain
     // comprises mostly RecvPacket and Ack msgs.
-    src_operational_data: Queue<OperationalData>,
-    dst_operational_data: Queue<OperationalData>,
+    pub(crate) src_operational_data: Queue<OperationalData>,
+    pub(crate) dst_operational_data: Queue<OperationalData>,
 
     // Toggle for the transaction confirmation mechanism.
     confirm_txes: bool,
@@ -197,13 +195,13 @@ impl<ChainA: ChainHandle, ChainB: ChainHandle> RelayPath<ChainA, ChainB> {
             .map_err(|e| LinkError::signer(self.dst_chain().id(), e))
     }
 
-    pub fn src_latest_height(&self) -> Result<Height, LinkError> {
+    pub(crate) fn src_latest_height(&self) -> Result<Height, LinkError> {
         self.src_chain()
             .query_latest_height()
             .map_err(|e| LinkError::query(self.src_chain().id(), e))
     }
 
-    pub fn dst_latest_height(&self) -> Result<Height, LinkError> {
+    pub(crate) fn dst_latest_height(&self) -> Result<Height, LinkError> {
         self.dst_chain()
             .query_latest_height()
             .map_err(|e| LinkError::query(self.dst_chain().id(), e))
@@ -217,15 +215,15 @@ impl<ChainA: ChainHandle, ChainB: ChainHandle> RelayPath<ChainA, ChainB> {
         Self::chain_time_at_height(self.dst_chain(), height)
     }
 
-    fn src_time_latest(&self) -> Result<Instant, LinkError> {
+    pub(crate) fn src_time_latest(&self) -> Result<Instant, LinkError> {
         self.src_time_at_height(Height::zero())
     }
 
-    fn dst_time_latest(&self) -> Result<Instant, LinkError> {
+    pub(crate) fn dst_time_latest(&self) -> Result<Instant, LinkError> {
         self.dst_time_at_height(Height::zero())
     }
 
-    fn src_max_block_time(&self) -> Result<Duration, LinkError> {
+    pub(crate) fn src_max_block_time(&self) -> Result<Duration, LinkError> {
         // TODO(hu55a1n1): Ideally, we should get the `max_expected_time_per_block` using the
         // `/genesis` endpoint once it is working in tendermint-rs.
         Ok(self
@@ -235,7 +233,7 @@ impl<ChainA: ChainHandle, ChainB: ChainHandle> RelayPath<ChainA, ChainB> {
             .max_block_time)
     }
 
-    fn dst_max_block_time(&self) -> Result<Duration, LinkError> {
+    pub(crate) fn dst_max_block_time(&self) -> Result<Duration, LinkError> {
         Ok(self
             .dst_chain()
             .config()
@@ -1620,91 +1618,6 @@ impl<ChainA: ChainHandle, ChainB: ChainHandle> RelayPath<ChainA, ChainB> {
         self.src_operational_data.replace(unelapsed_src_ods);
         self.dst_operational_data.replace(unelapsed_dst_ods);
         Ok((elapsed_src_ods, elapsed_dst_ods))
-    }
-
-    fn wait_for_conn_delay<ChainTime, MaxBlockTime, LatestHeight>(
-        odata: OperationalData,
-        chain_time: &ChainTime,
-        max_expected_time_per_block: &MaxBlockTime,
-        latest_height: &LatestHeight,
-    ) -> Result<OperationalData, LinkError>
-    where
-        ChainTime: Fn() -> Result<Instant, LinkError>,
-        MaxBlockTime: Fn() -> Result<Duration, LinkError>,
-        LatestHeight: Fn() -> Result<Height, LinkError>,
-    {
-        let (time_left, blocks_left) =
-            odata.conn_delay_remaining(chain_time, max_expected_time_per_block, latest_height)?;
-
-        match (time_left, blocks_left) {
-            (Duration::ZERO, 0) => {
-                info!(
-                    "ready to fetch a scheduled op. data with batch of size {} targeting {}",
-                    odata.batch.len(),
-                    odata.target,
-                );
-                Ok(odata)
-            }
-            (Duration::ZERO, blocks_left) => {
-                info!(
-                    "waiting ({:?} blocks left) for a scheduled op. data with batch of size {} targeting {}",
-                    blocks_left,
-                    odata.batch.len(),
-                    odata.target,
-                );
-
-                let blocks_left: u32 = blocks_left.try_into().expect("blocks_left > u32::MAX");
-
-                // Wait until the delay period passes
-                thread::sleep(blocks_left * max_expected_time_per_block()?);
-
-                Ok(odata)
-            }
-            (time_left, _) => {
-                info!(
-                    "waiting ({:?} left) for a scheduled op. data with batch of size {} targeting {}",
-                    time_left,
-                    odata.batch.len(),
-                    odata.target,
-                );
-
-                // Wait until the delay period passes
-                thread::sleep(time_left);
-
-                // `blocks_left` maybe non-zero, so recurse to recheck that all delays are handled.
-                Self::wait_for_conn_delay(
-                    odata,
-                    chain_time,
-                    max_expected_time_per_block,
-                    latest_height,
-                )
-            }
-        }
-    }
-
-    /// Fetches an operational data that has fulfilled its predefined delay period. May _block_
-    /// waiting for the delay period to pass.
-    /// Returns `Ok(None)` if there is no operational data scheduled.
-    pub(crate) fn fetch_scheduled_operational_data(
-        &self,
-    ) -> Result<Option<OperationalData>, LinkError> {
-        if let Some(odata) = self.src_operational_data.pop_front() {
-            Ok(Some(Self::wait_for_conn_delay(
-                odata,
-                &|| self.src_time_latest(),
-                &|| self.src_max_block_time(),
-                &|| self.src_latest_height(),
-            )?))
-        } else if let Some(odata) = self.dst_operational_data.pop_front() {
-            Ok(Some(Self::wait_for_conn_delay(
-                odata,
-                &|| self.dst_time_latest(),
-                &|| self.dst_max_block_time(),
-                &|| self.dst_latest_height(),
-            )?))
-        } else {
-            Ok(None)
-        }
     }
 
     fn restore_src_client(&self) -> ForeignClient<ChainA, ChainB> {
